Rework the spec file for CMake and follow openSUSE Packaging Guidelines
With the change from QMake to CMake, the spec needed adjusting to
build with CMake. In addition, the spec was not quite following the
openSUSE Packaging Guidelines exactly. In particular, the library
was not split out into its own subpackage and the license files were
not installed as license files.

Now, the packaging complies with the guidelines and enables other
packages to depend on the library without pulling in the whole
application.